feat(dashboard): audit async states, split PortfolioDashboard, add loading.tsx coverage
- Replace bespoke error/empty in PortfolioDashboard with shared DataBoundary
  and EmptyState primitives (Issue 1)
- Confirm async-states dev-only gate: notFound() guard + devOnly:true in
  routeMetadata excludes it from nav and command palette; add doc comment (Issue 2)
- Split PortfolioDashboard into SummarySection, AllocationSection, and
  ActivitySection subcomponents (Issue 3)
- Add missing loading.tsx for settings/notifications, settings/preferences,
  and settings/security using SettingsSectionSkeleton pattern (Issue 4)
- Update .gitignore: add test snapshots, Thumbs.db, editor directories
